chore: Move ConversationCallSystemMessageCellDescription into its own file
