+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/Components/ConversationSystemMessageCell.swift
@@ -194,48 +194,6 @@ private extension ConversationLike {
 
 // MARK: - Descriptions
 
-class ConversationCallSystemMessageCellDescription: ConversationMessageCellDescription {
-    typealias View = ConversationSystemMessageCell
-    let configuration: View.Configuration
-
-    var message: ZMConversationMessage?
-    weak var delegate: ConversationMessageCellDelegate?
-    weak var actionController: ConversationMessageActionController?
-
-    var showEphemeralTimer: Bool = false
-    var topMargin: Float = 0
-
-    let isFullWidth: Bool = true
-    let supportsActions: Bool = false
-    let containsHighlightableContent: Bool = false
-
-    let accessibilityIdentifier: String? = nil
-    let accessibilityLabel: String?
-
-    init(message: ZMConversationMessage, data: ZMSystemMessageData, missed: Bool) {
-        let viewModel = CallCellViewModel(
-            icon: missed ? .endCall : .phone,
-            iconColor: missed ? IconColors.backgroundMissedPhoneCall : IconColors.backgroundPhoneCall,
-            systemMessageType: data.systemMessageType,
-            font: .mediumFont,
-            textColor: LabelColors.textDefault,
-            message: message
-        )
-
-        configuration = View.Configuration(icon: viewModel.image(), attributedText: viewModel.attributedTitle(), showLine: false)
-        accessibilityLabel = viewModel.attributedTitle()?.string
-        actionController = nil
-    }
-
-    func isConfigurationEqual(with other: Any) -> Bool {
-        guard let otherDescription = other as? ConversationCallSystemMessageCellDescription else {
-            return false
-        }
-
-        return self.configuration == otherDescription.configuration
-    }
-}
-
 class ConversationVerifiedSystemMessageSectionDescription: ConversationMessageCellDescription {
     typealias View = ConversationSystemMessageCell
     let configuration: View.Configuration
